Courage is your greatest present need.

+++ index.php
@@ -58,7 +58,7 @@ function process_video($fileinfo)
         $cmd = "ffmpeg -i $old_file_name $video_setting $audio_setting $new_file_name";
         $results = shell_exec($cmd);
     }
-    if (!($same_video || $same_audio || $codecs['container'] == 'mp4')) {
+    if (!($same_video || $same_audio || $codecs['container'] == 'mp4') && $delete_on_conversion) {
         unlink(escapeshellarg($fileinfo->getPathname()));
     }
     print_r($results);
